requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
pillow>=8.3.0
tqdm>=4.62.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.2.0
hydra-core>=1.2.0

# Evaluation metrics
torchmetrics>=0.9.0
lpips>=0.1.4
clean-fid>=0.1.35

# Data augmentation
albumentations>=1.2.0

# Optional dependencies for advanced features
# accelerate>=0.20.0  # For distributed training
# wandb>=0.13.0        # For experiment tracking
# tensorboard>=2.10.0  # For logging
# streamlit>=1.20.0    # For demo app
# gradio>=3.30.0       # Alternative demo app

# Development dependencies
black>=22.0.0
ruff>=0.0.250
pytest>=7.0.0
pre-commit>=2.20.0
